# openapi.yaml
openapi: 3.0.3
info:
  title: StudySync API
  description: Backend API for the StudySync collaborative learning platform. Reliability score endpoints are not included.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDisplayProfile:
      type: object
      properties:
        fullName:
          type: string
        studentRole:
          type: string
        primaryUniversity:
          type: string
        secondaryUniversity:
          type: string
        email:
          type: string
        location:
          type: string
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /users/me/profile:
    get:
      tags: [Users]
      summary: Get display profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDisplayProfile'
    put:
      tags: [Users]
      summary: Update display profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDisplayProfile'
      responses:
        '200':
          description: Updated
  /users/me/avatar:
    post:
      tags: [Users]
      summary: Upload profile photo (multipart field photo, max 5MB)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Returns avatarUrl
    delete:
      tags: [Users]
      summary: Remove profile photo
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Removed
  /users/{userId}/avatar:
    get:
      tags: [Users]
      summary: Stream profile photo (self or shared group member)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image stream
  /users/me/groups:
    get:
      tags: [Users]
      summary: List user's study groups (dashboard)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /users/{userId}/profile:
    get:
      tags: [Users]
      summary: View another member's profile (shared group required)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public pod profile (no email for other users)
        '403':
          description: Not in a shared study group
...
  /workspaces/{groupId}:
    get:
      tags: [Workspaces]
      summary: Workspace metadata and members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /users/me/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: groupId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Notification list with unreadCount and nextCursor
  /users/me/notifications/read-all:
    post:
      tags: [Notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated count
  /users/me/notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Unread notification count
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /users/me/notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Mark one notification read or unread
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated notification
...
  /workspaces/{groupId}/files:
    get:
      tags: [Workspaces]
      summary: List shared files
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      tags: [Workspaces]
      summary: Upload shared file
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded
...